refactor(mouse): split the drag-selection clusters into mouse/selection.rs (F4 Part B, 2/n)
Verbatim relocation of the four `begin`/`extend`/`finish` triples —
pager text, chrome, pane text, list rows — out of `mouse/mod.rs`.
398 lines.

The boundary held up better than step 1 suggested it might. Only twelve
methods needed widening, and only to `pub(super)`: the triples
`handle_mouse` dispatches. The three helpers behind them
(`chrome_col_at`, `pane_cell_at`, `list_row_at`) stayed private, which
is the shape of a real module — a small entry surface over a private
interior — rather than a line-count-driven cut.

Everything else was path depth: `super::effect::` meant `app::effect`
from `mouse.rs` and needed one more `super::` from a level down, plus
moving the `Rect` and `clamp_to_area` imports to where they're now used.
No new `pub` fields, unlike step 1's `AgentViewInputs`.

Seams checked before cutting this time, not after: step 1's boundary
landed mid-doc-comment and only clippy noticed.

mod.rs 2,538 → 2,156 lines. scroll.rs and forward.rs follow; the
remaining production half is `handle_mouse` plus the data types.
